Use signed ints

+++ map.c
@@ -11,8 +11,8 @@ static pair * get_pair(bucket *bucket, const char *key) {
   return NULL;
 }
 
-static uint64_t hash(const char *str) {
-  uint64_t hash = 5381;
+static int hash(const char *str) {
+  int hash = 5381;
   char c;
 
   while (c = *str++) {
@@ -21,14 +21,14 @@ static uint64_t hash(const char *str) {
   return hash;
 }
 
-void map_init(map *map, const uint64_t capacity) {
+void map_init(map *map, const int capacity) {
   map->count = capacity;
   map->buckets = malloc(map->count * sizeof(bucket));
   memset(map->buckets, 0, map->count * sizeof(bucket));
 }
 
-void map_set(map *map, char *key, uint64_t value) {
-  uint64_t index;
+void map_set(map *map, char *key, int value) {
+  int index;
   bucket *bucket;
   pair *pair;
 
@@ -55,7 +55,7 @@ void map_set(map *map, char *key, uint64_t value) {
   }
 }
 
-uint64_t map_get(map *map, char *key) {
+int map_get(map *map, char *key) {
   size_t index;
   bucket *bucket;
   pair *pair;
